UPDATE: Moved pagination's meta dto to its own module
